backend/requirements.txt
django>=4.2,<5.0
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-filter
django-storages
whitenoise
gunicorn
uvicorn[standard]
psycopg2-binary
celery
redis
channels[daphne]
numpy==1.26.4
opencv-python-headless==4.9.0.80
librosa
scipy
pandas
scikit-learn
torch
torchvision
torchaudio
transformers
Pillow
onnxruntime
web3
python-dotenv
drf-spectacular
